feat : 예약 가능 확인 GET

+++ src/main/java/Project/PENBOT/Booking/Repository/BookingRepository.java
@@ -4,8 +4,9 @@ import Project.PENBOT.Booking.Entity.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDate;
+import java.util.List;
 
-public interface BookingRepository extends JpaRepository<Booking, Long> {
+public interface BookingRepository extends JpaRepository<Booking, Integer> {
     /**
      *  겹치는 예약이 있는지 조회
      * */
